# lib/api-zod/src/generated/types/actionProposal.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionEvent } from './actionEvent';
import type { ActionProposalApproval } from './actionProposalApproval';
import type { ActionProposalCounts } from './actionProposalCounts';
import type { ActionProposalEvidence } from './actionProposalEvidence';
import type { ActionProposalExecution } from './actionProposalExecution';
import type { ActionProposalPostflight } from './actionProposalPostflight';
import type { ActionProposalPreflight } from './actionProposalPreflight';
import type { ActionProposalRevert } from './actionProposalRevert';
import type { ActionProposalStatus } from './actionProposalStatus';
import type { ActionProposalVerification } from './actionProposalVerification';
import type { ActionRisk } from './actionRisk';
import type { ActionSource } from './actionSource';
import type { ActionStep } from './actionStep';
import type { ActionTarget } from './actionTarget';
import type { ActionType } from './actionType';

export interface ActionProposal {
  id: number;
  proposalKey: string;
  type: ActionType;
  source: ActionSource;
  reason: string;
  status: ActionProposalStatus;
  risk: ActionRisk;
  requiresApproval: boolean;
  dryRun: boolean;
  allowCreateDirectories: boolean;
  /** @nullable */
  reviewItemId: number | null;
  /** @nullable */
  acquisitionJobId: number | null;
  /** @nullable */
  downloadJobId: number | null;
  planHash: string;
  target: ActionTarget;
  evidence: ActionProposalEvidence;
  approval: ActionProposalApproval;
  preflight: ActionProposalPreflight;
  execution: ActionProposalExecution;
  verification: ActionProposalVerification;
  revert: ActionProposalRevert;
  postflight: ActionProposalPostflight;
  retryCount: number;
  maxRetries: number;
  /** @nullable */
  errorCode: string | null;
  /** @nullable */
  errorMessage: string | null;
  counts: ActionProposalCounts;
  steps: ActionStep[];
  events: ActionEvent[];
  createdAt: Date;
  /** @nullable */
  approvedAt: Date | null;
  /** @nullable */
  executedAt: Date | null;
  /** @nullable */
  completedAt: Date | null;
  /** @nullable */
  cancelledAt: Date | null;
  updatedAt: Date;
}