# lib/api-zod/src/generated/types/actionProposalStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ActionProposalStatus = typeof ActionProposalStatus[keyof typeof ActionProposalStatus];


export const ActionProposalStatus = {
  draft: 'draft',
  proposed: 'proposed',
  approved: 'approved',
  preflight: 'preflight',
  ready: 'ready',
  executing: 'executing',
  completed: 'completed',
  partially_completed: 'partially_completed',
  failed: 'failed',
  cancelled: 'cancelled',
  reverted: 'reverted',
} as const;