# lib/api-zod/src/generated/types/actionStepSelection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionStepSelectionSelectionsItem } from './actionStepSelectionSelectionsItem';

export interface ActionStepSelection {
  /** @minItems 1 */
  selections: ActionStepSelectionSelectionsItem[];
}