# lib/api-zod/src/generated/types/actionStepStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ActionStepStatus = typeof ActionStepStatus[keyof typeof ActionStepStatus];


export const ActionStepStatus = {
  pending: 'pending',
  skipped: 'skipped',
  ready: 'ready',
  executing: 'executing',
  completed: 'completed',
  failed: 'failed',
  reverted: 'reverted',
} as const;