# lib/api-zod/src/generated/types/namingActionCandidate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionType } from './actionType';

export interface NamingActionCandidate {
  fileRecordId: number;
  type: ActionType;
  sourcePath: string;
  destinationPath: string;
  sourceFilename: string;
  destinationFilename: string;
  confidence: string;
  patternId: string;
  reason: string;
  evidence: string[];
}