# lib/api-zod/src/generated/types/namingActionCandidates.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NamingActionCandidate } from './namingActionCandidate';
import type { NamingActionCandidatesSummary } from './namingActionCandidatesSummary';

export interface NamingActionCandidates {
  summary: NamingActionCandidatesSummary;
  candidates: NamingActionCandidate[];
}