# lib/api-zod/src/generated/types/planNamingNormalization.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlanNamingNormalization {
  confidence?: string;
  pattern?: string;
  mediaType?: string;
  volume?: string;
  /**
     * @minimum 1
     * @maximum 500
     */
  limit?: number;
  /** @items.minimum 1 */
  fileRecordIds?: number[];
}